+++ RoomInn/app/src/main/java/postpc/finalproject/RoomInn/furnitureData/Door.kt
@@ -2,9 +2,7 @@ package postpc.finalproject.RoomInn.furnitureData
 
 import android.graphics.Color
 import android.graphics.Path
-import android.util.Size
 import postpc.finalproject.RoomInn.R
-import java.util.*
 
 class Door(
     position: Point3D = Point3D(),
@@ -47,84 +45,84 @@ class Door(
     }
 
 
-    override fun draw(sizeWidth: Float, sizeHeight: Float): Path {
+    override fun draw(sizeWidth: Double, sizeHeight: Double): Path {
         val path = Path()
         val margin = 8f
         if (unityType.typeName == "Right") {
-            path.moveTo(margin, ((scale.x * sizeHeight) + margin))
+            path.moveTo(margin, (((scale.x * sizeHeight) + margin).toFloat()))
             path.arcTo(
-                -scale.x * sizeWidth + margin,
+                (-scale.x * sizeWidth + margin).toFloat(),
                 margin,
-                scale.x * sizeWidth + margin,
-                (scale.x * sizeHeight * 2) + margin,
+                (scale.x * sizeWidth + margin).toFloat(),
+                ((scale.x * sizeHeight * 2) + margin).toFloat(),
                 -90f, 90f, false
             )
 
             path.lineTo(
-                margin * 0.5f + ((scale.x * sizeWidth) * 0.9f),
-                ((scale.x * sizeHeight) + margin)
+                (margin * 0.5f + ((scale.x * sizeWidth) * 0.9f)).toFloat(),
+                (((scale.x * sizeHeight) + margin).toFloat())
             )
             path.moveTo(
-                margin * 0.5f + ((scale.x * sizeWidth) * 0.75f),
-                ((scale.x * sizeHeight) + margin)
+                (margin * 0.5f + ((scale.x * sizeWidth) * 0.75f)).toFloat(),
+                (((scale.x * sizeHeight) + margin).toFloat())
             )
             path.lineTo(
-                margin * 0.5f + ((scale.x * sizeWidth) * 0.6f),
-                ((scale.x * sizeHeight) + margin)
+                (margin * 0.5f + ((scale.x * sizeWidth) * 0.6f)).toFloat(),
+                (((scale.x * sizeHeight) + margin).toFloat())
             )
             path.moveTo(
-                margin * 0.5f + ((scale.x * sizeWidth) * 0.45f),
-                ((scale.x * sizeHeight) + margin)
+                (margin * 0.5f + ((scale.x * sizeWidth) * 0.45f)).toFloat(),
+                (((scale.x * sizeHeight) + margin).toFloat())
             )
             path.lineTo(
-                margin * 0.5f + ((scale.x * sizeWidth) * 0.3f),
-                ((scale.x * sizeHeight) + margin)
+                (margin * 0.5f + ((scale.x * sizeWidth) * 0.3f)).toFloat(),
+                (((scale.x * sizeHeight) + margin).toFloat())
             )
             path.moveTo(
-                margin * 0.5f + ((scale.x * sizeWidth) * 0.15f),
-                ((scale.x * sizeHeight) + margin)
+                (margin * 0.5f + ((scale.x * sizeWidth) * 0.15f)).toFloat(),
+                (((scale.x * sizeHeight) + margin).toFloat())
             )
-            path.lineTo(margin * 0.5f, ((scale.x * sizeHeight) + margin))
+            path.lineTo(margin * 0.5f, (((scale.x * sizeHeight) + margin).toFloat()))
 
-            path.moveTo(margin * 2, ((scale.x * sizeHeight) + margin))
+            path.moveTo(margin * 2, (((scale.x * sizeHeight) + margin).toFloat()))
             path.lineTo(margin * 2, margin)
         } else {
             path.moveTo(margin , margin)
 
             path.arcTo(
-                (-scale.x * sizeWidth+margin),
-                (-scale.x * sizeHeight+margin),
-                (scale.x * sizeWidth+margin),
-                (scale.x * sizeHeight+margin),
+                ((-scale.x * sizeWidth+margin).toFloat()),
+                ((-scale.x * sizeHeight+margin).toFloat()),
+                ((scale.x * sizeWidth+margin).toFloat()),
+                ((scale.x * sizeHeight+margin).toFloat()),
                 90f, -90f, false
             )
             path.lineTo(
-                margin * 0.5f + ((scale.x * sizeWidth) * 0.9f),
+                (margin * 0.5f + ((scale.x * sizeWidth) * 0.9f)).toFloat(),
                 (margin)
             )
             path.moveTo(
-                margin * 0.5f + ((scale.x * sizeWidth) * 0.75f),
+                (margin * 0.5f + ((scale.x * sizeWidth) * 0.75f)).toFloat(),
                 (margin)
             )
             path.lineTo(
-                margin * 0.5f + ((scale.x * sizeWidth) * 0.6f),
+                (margin * 0.5f + ((scale.x * sizeWidth) * 0.6f)).toFloat(),
                 (margin)
             )
             path.moveTo(
-                margin * 0.5f + ((scale.x * sizeWidth) * 0.45f),
+                (margin * 0.5f + ((scale.x * sizeWidth) * 0.45f)).toFloat(),
                 (margin)
             )
             path.lineTo(
-                margin * 0.5f + ((scale.x * sizeWidth) * 0.3f),
+                (margin * 0.5f + ((scale.x * sizeWidth) * 0.3f)).toFloat(),
                 (margin)
             )
             path.moveTo(
-                margin * 0.5f + ((scale.x * sizeWidth) * 0.15f),
+                (margin * 0.5f + ((scale.x * sizeWidth) * 0.15f)).toFloat(),
                 (margin)
             )
             path.lineTo(margin * 0.5f, (margin))
 
-            path.moveTo(margin * 2, ((scale.x * sizeHeight) + margin))
+            path.moveTo(margin * 2, (((scale.x * sizeHeight) + margin).toFloat()))
             path.lineTo(margin * 2, margin)
         }
     return path
@@ -132,7 +130,7 @@ class Door(
 
     override fun unityScale(): Point3D {
         return scale.getDivideByPoint(unityType.defaultScale).apply {
-            z = 1f
+            z = 1.0
         }
     }
